+++ CHANGELOG.md
@@ -31,7 +31,8 @@
 ### Build
 
  * Restructure `progress.json` file
- * Use sha-512 hash to create build 'descriptions' and 'reports' filenames
+ * Put dataset 'Description' into 'Report'
+ * Use sha-512 hash to create 'reports' filenames
 
 ### Development